webhook-server/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
httpx==0.25.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.4.2